More compiler warnings...
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@2833 7a7537e8-13f0-0310-91df-b6672ffda945

+++ scheduler/auth.c
@@ -1,5 +1,5 @@
 /*
- * "$Id: auth.c,v 1.62 2002/09/26 11:31:05 mike Exp $"
+ * "$Id: auth.c,v 1.63 2002/09/26 15:19:30 mike Exp $"
  *
  *   Authorization routines for the Common UNIX Printing System (CUPS).
  *
@@ -950,7 +950,7 @@ IsAuthorized(client_t *con)	/* I - Connection */
     return (HTTP_OK);
 
   LogMessage(L_DEBUG2, "IsAuthorized: username = \"%s\" password = %d chars",
-	     con->username, strlen(con->password));
+	     con->username, (int)strlen(con->password));
   DEBUG_printf(("IsAuthorized: username = \"%s\", password = \"%s\"\n",
 		con->username, con->password));
 
@@ -1623,5 +1623,5 @@ to64(char          *s,	/* O - Output string */
 
 
 /*
- * End of "$Id: auth.c,v 1.62 2002/09/26 11:31:05 mike Exp $".
+ * End of "$Id: auth.c,v 1.63 2002/09/26 15:19:30 mike Exp $".
  */
